move daily common to its own file
