+++ release/pipeline/dags/istio_common_dag.py
@@ -235,146 +235,3 @@ def MakeCommonDag(dag_args_func, name,
   tasks['copy_files_for_release'] = copy_files
 
   return common_dag, tasks
-
-
-def ReportDailySuccessful(task_instance, **kwargs):
-  """Set this release as the candidate if it is the latest."""
-  date = kwargs['execution_date']
-  branch = GetSettingPython(task_instance, 'BRANCH')
-  latest_run = float(GetVariableOrDefault(branch+'latest_daily_timestamp', 0))
-
-  timestamp = time.mktime(date.timetuple())
-  logging.info("Current run's timestamp: %s \n"
-               "latest_daily's timestamp: %s", timestamp, latest_run)
-  if timestamp >= latest_run:
-    run_sha = task_instance.xcom_pull(task_ids='get_git_commit')
-    latest_version = GetSettingPython(task_instance, 'VERSION')
-
-    Variable.set(branch+'_latest_sha', run_sha)
-    Variable.set(branch+'_latest_daily', latest_version)
-    Variable.set(branch+'_latest_daily_timestamp', timestamp)
-
-    logging.info('%s_latest_sha set to %s', branch, run_sha)
-    logging.info('setting latest green daily of %s branch to: %s', branch, run_sha)
-    return 'tag_daily_gcr'
-  return 'skip_tag_daily_gcr'
-
-
-def MakeMarkComplete(dag):
-  """Make the final sequence of the daily graph."""
-  mark_complete = BranchPythonOperator(
-      task_id='mark_complete',
-      python_callable=ReportDailySuccessful,
-      provide_context=True,
-      dag=dag,
-  )
-
-  gcr_tag_success = r"""
-{% set settings = task_instance.xcom_pull(task_ids='generate_workflow_args') %}
-set -x
-pwd; ls
-
-gsutil ls gs://{{ settings.GCS_FULL_STAGING_PATH }}/docker/           > docker_tars.txt
-cat docker_tars.txt |   grep -Eo "docker\/(([a-z]|[0-9]|-|_)*).tar.gz" | \
-                        sed -E "s/docker\/(([a-z]|[0-9]|-|_)*).tar.gz/\1/g" > docker_images.txt
-
-  gcloud auth configure-docker  -q
-  cat docker_images.txt | \
-  while read -r docker_image;do
-    gcloud container images add-tag \
-    "gcr.io/{{ settings.GCR_STAGING_DEST }}/${docker_image}:{{ settings.VERSION }}" \
-    "gcr.io/{{ settings.GCR_STAGING_DEST }}/${docker_image}:{{ settings.BRANCH }}-latest-daily" --quiet;
-    #pull_source="gcr.io/{{ settings.GCR_STAGING_DEST }}/${docker_image}:{{ settings.VERSION }}"
-    #push_dest="  gcr.io/{{ settings.GCR_STAGING_DEST }}/${docker_image}:latest_{{ settings.BRANCH }}";
-    #docker pull $pull_source
-    #docker tag  $pull_source $push_dest
-    #docker push $push_dest
-  done
-
-cat docker_tars.txt docker_images.txt
-rm  docker_tars.txt docker_images.txt
-"""
-
-  tag_daily_grc = BashOperator(
-      task_id='tag_daily_gcr',
-      bash_command=gcr_tag_success,
-      dag=dag,
-  )
-  # skip_grc = DummyOperator(
-  #     task_id='skip_tag_daily_gcr',
-  #     dag=dag,
-  # )
-  # end = DummyOperator(
-  #     task_id='end',
-  #     dag=dag,
-  #     trigger_rule="one_success",
-  # )
-  mark_complete >> tag_daily_grc
-  # mark_complete >> skip_grc >> end
-  return mark_complete
-
-
-
-def DailyPipeline(branch):
-  def DailyGenerateTestArgs(**kwargs):
-    daily_params = dict(
-      GCS_DAILY_PATH='daily-build/{version}',
-      MFEST_COMMIT='{branch}@{{{timestamp}}}',
-      VERIFY_CONSISTENCY='false',
-      VERSION='{branch}-{date_string}')
-
-    """Loads the configuration that will be used for this Iteration."""
-    conf = kwargs['dag_run'].conf
-    if conf is None:
-      conf = dict()
-
-    # If variables are overriden then we should use it otherwise we use it's
-    # default value.
-    date = datetime.datetime.now()
-    date_string = date.strftime('%Y%m%d-%H-%M')
-
-    version = conf.get('VERSION')
-    if version is None:
-      version = daily_params['VERSION'].format(
-        branch=branch, date_string=date_string)
-
-    gcs_path = conf.get('GCS_DAILY_PATH')
-    if gcs_path is None:
-       gcs_path = daily_params['GCS_DAILY_PATH'].format(
-                      version=version)
-
-    mfest_commit = conf.get('MFEST_COMMIT')
-    if mfest_commit is None:
-       timestamp = time.mktime(date.timetuple())
-       mfest_commit = daily_params['MFEST_COMMIT'].format(
-          branch=branch, timestamp=timestamp)
-
-    default_conf = environment_config.get_default_config(
-        branch=branch,
-        gcs_path=gcs_path,
-        mfest_commit=mfest_commit,
-        verify_consistency=daily_params['VERIFY_CONSISTENCY'],
-        version=version)
-
-    config_settings = dict()
-    for name in default_conf.iterkeys():
-      config_settings[name] = conf.get(name) or default_conf[name]
-    config_settings['PIPELINE_TYPE'] = 'daily'
-    return config_settings
-
-  dag_name = 'istio_daily_' + branch
-  dag, tasks = MakeCommonDag(
-       DailyGenerateTestArgs,
-       name=dag_name, schedule_interval='15 9 * * *')
-  tasks['mark_daily_complete'] = MakeMarkComplete(dag)
-
-  #tasks['generate_workflow_args']
-  tasks['get_git_commit'                 ].set_upstream(tasks['generate_workflow_args'])
-  tasks['run_cloud_builder'              ].set_upstream(tasks['get_git_commit'])
-  tasks['run_release_qualification_tests'].set_upstream(tasks['run_cloud_builder'])
-  tasks['modify_values_helm'             ].set_upstream(tasks['run_release_qualification_tests'])
-  tasks['copy_files_for_release'         ].set_upstream(tasks['modify_values_helm'])
-  tasks['mark_daily_complete'            ].set_upstream(tasks['copy_files_for_release'])
-
-  return dag
-
